+++ src/index.ts
@@ -1,20 +1,13 @@
 export * from './components/Dashboard';
-export { default as Dashboard } from './components/Dashboard';
 
 export * from './components/ListForm';
-export { default as ListForm } from './components/ListForm';
 
 export * from './components/ListItemPicker';
-export { default as ListItemPicker } from './components/ListItemPicker';
 
 export * from './components/Navigation';
-export { default as Navigation } from './components/Navigation';
 
 export * from './components/PeoplePicker';
-export { default as PeoplePicker } from './components/PeoplePicker';
 
 export * from './components/Searchbar';
-export { default as Searchbar } from './components/Searchbar';
 
 export * from './components/SiteBreadcrumb';
-export { default as SiteBreadcrumb } from './components/SiteBreadcrumb';