+++ CHANGELOG.md
@@ -4,9 +4,8 @@
 
 ### **Bug Fixes & Improvements**
 
-- **Fixed an import bug**
-  - Fixed issue with Breadcrumb
-  - New ListForm props added
+- Fixed issue with Breadcrumb
+- New ListForm props added
 
 ## **v0.6.3** (Bug Fix Release)